added biting
also converted ints in damage calculation to doubles and cleaned up the main interface class

+++ RSHitChance2/src/abilitiesSpecific/BleedAbility.java
@@ -11,12 +11,12 @@ import abilities.BaseAbility;
 import abilities.PercentageRange;
 import abilities.PercentageRangeFlyweight;
 import buff.Buff;
+import calculations.AbilityHit;
+import calculations.BleedHitSlantedAverage;
+import calculations.Hit;
 import combatStyle.OffensiveCombatStyle;
 import combatent.Combatent;
-import main.AbilityHit;
-import main.BleedHitSlantedAverage;
 import main.DamageMode;
-import main.Hit;
 
 public class BleedAbility extends BaseAbility{
 
@@ -55,7 +55,7 @@ public class BleedAbility extends BaseAbility{
     }
 
     @Override
-    public List<Hit> generateBaseDamageHitList(int baseDamage, OffensiveCombatStyle combatStyle) {
+    public List<Hit> generateBaseDamageHitList(double baseDamage, OffensiveCombatStyle combatStyle) {
         List<Hit> hitList = new ArrayList<>(numberOfHits);
         for(int i=0;i<numberOfHits;i++) {
             hitList.add(new BleedHitSlantedAverage(combatStyle,baseDamage*range.getMin()/numberOfHits,baseDamage*range.getMax()/numberOfHits));
