+++ RSHitChance2/src/abilitiesSpecific/ChanneledAbility.java
@@ -11,12 +11,12 @@ import abilities.BaseAbility;
 import abilities.PercentageRange;
 import abilities.PercentageRangeFlyweight;
 import buff.Buff;
+import calculations.AbilityHit;
+import calculations.BleedHitSlantedAverage;
+import calculations.Hit;
 import combatStyle.OffensiveCombatStyle;
 import combatent.Combatent;
-import main.AbilityHit;
-import main.BleedHitSlantedAverage;
 import main.DamageMode;
-import main.Hit;
 
 public class ChanneledAbility extends BaseAbility{
 
@@ -47,7 +47,7 @@ public class ChanneledAbility extends BaseAbility{
     }
 
     @Override
-    public List<Hit> generateBaseDamageHitList(int baseDamage, OffensiveCombatStyle combatStyle) {
+    public List<Hit> generateBaseDamageHitList(double baseDamage, OffensiveCombatStyle combatStyle) {
         List<Hit> hitList = new ArrayList<>(ranges.length);
         for(int i=0;i<ranges.length;i++) {
             hitList.add(new AbilityHit(combatStyle,baseDamage*ranges[i].getMin(),baseDamage*ranges[i].getMax()));
