+++ RSHitChance2/src/abilitiesSpecific/MassacreClass.java
@@ -11,13 +11,13 @@ import abilities.BaseAbility;
 import abilities.PercentageRange;
 import abilities.PercentageRangeFlyweight;
 import buff.Buff;
+import calculations.AbilityHit;
+import calculations.BleedHitSlantedAverage;
+import calculations.ConstantHit;
+import calculations.Hit;
 import combatStyle.OffensiveCombatStyle;
 import combatent.Combatent;
-import main.AbilityHit;
-import main.BleedHitSlantedAverage;
-import main.ConstantHit;
 import main.DamageMode;
-import main.Hit;
 /*
  * does 6 hits. the last 5 are bleeds
  */
@@ -49,11 +49,11 @@ public class MassacreClass extends BaseAbility{
     }
 
     @Override
-    public List<Hit> generateBaseDamageHitList(int baseDamage, OffensiveCombatStyle combatStyle) {
+    public List<Hit> generateBaseDamageHitList(double baseDamage, OffensiveCombatStyle combatStyle) {
         List<Hit> hitList = new ArrayList<>(6);
         hitList.add(new AbilityHit(combatStyle,baseDamage*rangeForFirstHit.getMin(),baseDamage*rangeForFirstHit.getMax()));
         for(int i=0;i<5;i++) {
-            hitList.add(new ConstantHit(combatStyle,(int) (baseDamage*0.6266)));//no range for the bleeds
+            hitList.add(new ConstantHit(combatStyle,baseDamage*0.6266));//no range for the bleeds
         }
         return hitList;
     }
